pyproject.toml
[project]
name = "poly-bot"
version = "0.1.0"
dependencies = [
    "fastmcp",
    "py-clob-client",
    "ib_insync",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv",
    "requests",
    "aiohttp",
    "websockets",
    "numpy",
    "pandas",
    "pytest>=7.0",
]
requires-python = ">=3.9"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
